feat:#174 Split dialog components
