+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogContent/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogContent.component';
