+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogDescription/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogDescription.component';
