+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogFooter/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogFooter.component';
