+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogHeader.component';
