+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogOverlay/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogOverlay.component';
