+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogPortal/dialogPortal.component.tsx
@@ -1,0 +1,12 @@
+import * as DialogPrimitive from '@radix-ui/react-dialog';
+
+import { cn } from '../../../lib/utils';
+
+const DialogPortal = ({ className, children, ...props }: DialogPrimitive.DialogPortalProps) => (
+  <DialogPrimitive.Portal className={cn(className)} {...props}>
+    <div className="fixed inset-0 z-50 flex items-start justify-center sm:items-center">{children}</div>
+  </DialogPrimitive.Portal>
+);
+DialogPortal.displayName = DialogPrimitive.Portal.displayName;
+
+export { DialogPortal };
