+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogPortal/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogPortal.component';
