+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogTitle.component';
