+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogTrigger/dialogTrigger.component.tsx
@@ -1,0 +1,5 @@
+import * as DialogPrimitive from '@radix-ui/react-dialog';
+
+const DialogTrigger = DialogPrimitive.Trigger;
+
+export { DialogTrigger };
