+++ packages/webapp-libs/webapp-core/src/components/dialog/dialogTrigger/index.ts
@@ -1,0 +1,1 @@
+export * from './dialogTrigger.component';
