+++ packages/webapp-libs/webapp-core/src/components/dialog/index.ts
@@ -1,1 +1,9 @@
 export * from './dialog.component';
+export * from './dialogContent';
+export * from './dialogDescription';
+export * from './dialogFooter';
+export * from './dialogHeader';
+export * from './dialogOverlay';
+export * from './dialogPortal';
+export * from './dialogTitle';
+export * from './dialogTrigger';
